modify/#14: 데이터베이스 값 확인 절차 추가

+++ app/src/models/comment/commentService.js
@@ -11,34 +11,52 @@ class CommentService {
     const boardNo = this.params.board_no;
     const userNo = this.body.user_no;
     const comments = this.body.content;
-    try {
-      const response = await CommentRepository.addComments(
-        boardNo,
-        userNo,
-        comments
-      );
-      return response;
-    } catch (err) {
-      return {success: false};
+
+    let error = await CommentRepository.isBoardNO(boardNo);
+    if (!error[0][0]) {
+      return {
+        error: "Not Found",
+        message: "해당 게시글이 존재하지 않습니다.",
+        Statuscode: 404,
+      };
     }
+    error = await CommentRepository.isUserNO(userNo);
+    if (!error[0][0]) {
+      return {
+        error: "Not Found",
+        message: "해당 유저가 존재하지 않습니다.",
+        Statuscode: 404,
+      };
+    }
+
+    await CommentRepository.addComments(boardNo, userNo, comments);
+    return {Statuscode: 201};
   }
   async deleteComments() {
     const No = this.body.no;
-    try {
-      const response = await CommentRepository.deleteComments(No);
-      return response;
-    } catch (err) {
-      return {success: false};
+    let error = await CommentRepository.isNO(No);
+    if (!error[0][0]) {
+      return {
+        error: "Not Found",
+        message: "해당 댓글이 존재하지 않습니다.",
+        Statuscode: 404,
+      };
     }
+    await CommentRepository.deleteComments(No);
+    return {message: "댓글이 성공적으로 삭제됐습니다.", Status: 200};
   }
   async getComments() {
     const boardNo = this.body.board_no;
-    try {
-      const [rows, fields] = await CommentRepository.getComments(boardNo);
-      return rows;
-    } catch {
-      return {success: false};
+    let error = await CommentRepository.isBoardNO(boardNo);
+    if (!error[0][0]) {
+      return {
+        error: "Not Found",
+        message: "해당 게시글이 존재하지 않습니다.",
+        statuscode: 404,
+      };
     }
+    const [rows, fields] = await CommentRepository.getComments(boardNo);
+    return {Statuscode: 201, rows};
   }
 }
